+++ bag.h
@@ -16,7 +16,7 @@ char long_long_buffer[8];
 struct Record {
     char *file_path;
     off_t offset;
-    unsigned long long file_size;
+    long int file_size;
 };
 
 void create(char *dir, char *file);
@@ -25,7 +25,7 @@ void scan(char *file, char *pattern);
 
 void extract(char *file, char *dir, char *pattern);
 
-struct Record read_file(FILE *bag, long long offset);
+struct Record read_file(FILE *bag, off_t offset);
 
 void write_file(const char *path, const char *name, FILE *bag);
 
@@ -33,6 +33,6 @@ void write_dir(const char *dir, FILE *bag, const char *dir_rel);
 
 void read_buffer(FILE *file, char *buf, size_t buf_size);
 
-void copy_to_file(FILE *in, const char *path, long long length);
+void copy_to_file(FILE *in, const char *path, long int length);
 
 #endif //BAG_BAG_H
