+++ utils.h
@@ -9,11 +9,11 @@
 
 void put_unsigned_short_be(char buf[2], unsigned short val);
 
-void put_long_long_be(char buf[8], long long val);
+void put_long_long_be(char buf[8], long int val);
 
 unsigned short get_unsigned_short_be(char buf[2]);
 
-unsigned long long get_long_long_be(char buf[8]);
+long int get_long_long_be(char buf[8]);
 
 int make_path(char *file_path, mode_t mode);
